lesson26/requirements.txt
pytest==6.1.0
selenium==3.141.0
pytest-xdist==2.1.0
allure-pytest==2.8.22
requests==2.24.0
jsonschema==3.2.0
paramiko==2.7.2
